Fix errors and warnings in tests
ImportError: cannot import name 'config' from 'data'

DeprecationWarning: invalid escape sequence \s

HypothesisDeprecationWarning: In blacklist_categories='Cc,N', 'c' is not
a valid Unicode category. This will be an error in a future version.

HypothesisDeprecationWarning: Test: test_password took 303.48ms to run.
In future the default deadline setting will be 200ms, which will make
this an error. You can set deadline to an explicit value of e.g. 400 to
turn tests slower than this into an error, or you can set it to None to
disable this check entirely.